Hero section, js => sections (new folder), css => components (added css files), home.css (added css files directories), home.js (changed js files directories).
