+++ css/components/hero.css
@@ -1,0 +1,63 @@
+.left-col {
+    margin-top: 10rem;
+}
+.right-col {
+}
+.right-col .hero {
+    float: right;
+    width: 90%;
+}
+.intro1 {
+    font-size: 2rem;
+    font-weight: bold;
+}
+.intro2 {
+    font-size: 4rem;
+    font-weight: bold;
+}
+.intro3 {
+    font-size: 3rem;
+    font-weight: bold;
+}
+p {
+    display: flex;
+    align-items: center;
+    line-height: 1.5;
+}
+.hero-description {
+    margin: 3rem 0;
+    font-weight: bold;
+    font-size: 1.2rem;
+    color: gray;
+}
+.btn-download-cv {
+    width: 11rem;
+    height: 4rem;
+    border-radius: 10px;
+    border-style: none;
+    margin-left: 2rem;
+    font-weight: bold;
+    font-size: 0.9rem;
+    box-shadow: 0px 0px 10px #6c6767;
+    cursor: pointer;
+    color: var(--gray-dark);
+}
+.btn-hire-me {
+    width: 11rem;
+    height: 4rem;
+    border-radius: 10px;
+    font-weight: bold;
+    font-size: 0.9rem;
+    border-style: none;
+    box-shadow: 0px 0px 10px #6c6767;
+    background-image: linear-gradient(
+        45deg,
+        rgb(255, 0, 118) 0%,
+        rgb(89, 15, 183) 100%
+    );
+    color: rgb(255, 255, 255);
+    cursor: pointer;
+}
+.fa-download {
+    margin-left: 0.5rem;
+}
