+++ js/pages/home.js
@@ -1,18 +1,18 @@
-import { header } from '../components/header.js';
-import { hero } from '../components/hero.js';
-import { about } from '../components/about.js';
-import { whatWeDo } from '../components/what-we-do.js';
-import { getInTouch } from '../components/get-in-touch.js';
-import { portfolio } from '../components/portfolio.js';
-import { pricing } from '../components/pricing.js';
-import { horizontalLineAfterPricing } from '../components/horizontal-line-after-pricing.js';
-import { testimonial } from '../components/testimonial.js';
-import { contactMe } from '../components/contact-me.js';
-import { blog } from '../components/blog.js';
-import { heroSocialMedia } from '../components/hero-social-media.js';
-import { horizontalLineAndFooter } from '../components/horizontal-line-and-footer.js';
-import { backToTopButton } from '../components/back-to-top-button.js';
-import { footerSocialMedia } from '../components/footer-social-media.js';
+import { header } from '../sections/header.js';
+import { hero } from '../sections/hero.js';
+import { about } from '../sections/about.js';
+import { whatWeDo } from '../sections/what-we-do.js';
+import { getInTouch } from '../sections/get-in-touch.js';
+import { portfolio } from '../sections/portfolio.js';
+import { pricing } from '../sections/pricing.js';
+import { horizontalLineAfterPricing } from '../sections/horizontal-line-after-pricing.js';
+import { testimonial } from '../sections/testimonial.js';
+import { contactMe } from '../sections/contact-me.js';
+import { blog } from '../sections/blog.js';
+import { heroSocialMedia } from '../sections/hero-social-media.js';
+import { horizontalLineAndFooter } from '../sections/horizontal-line-and-footer.js';
+import { backToTopButton } from '../sections/back-to-top-button.js';
+import { footerSocialMedia } from '../sections/footer-social-media.js';
 // Header
 header();
 // Hero
