+++ js/sections/horizontal-line-after-pricing.js
