+++ js/sections/horizontal-line-and-footer.js
